Rewrite my Neovim config
Take the modular kickstart config and rewrite to avoid some of the
complexity. The goals was to resolve a number of errors that had crept
in to some of the plugins. That seems to be mostly achieved. There are
still some deprecation warnings that I assume are from some plugins.